requirements.txt
mlagents-envs==0.9.1
tensorflow==2.0.0-beta1
tensorboard==1.14.0
numpy==1.14.5
protobuf
gym
tqdm
google-cloud-bigtable
google-cloud-storage